scripts/requirements.txt
qrcode[pil]
pillow
python-dotenv
gspread
google-auth
